requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.3
sentence-transformers>=2.7.0
torch>=2.2; platform_system != "Darwin" or platform_machine != "arm64"
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
tqdm>=4.66
rich>=13.7
openpyxl>=3.1
openai>=1.50.0
langchain-text-splitters>=0.2.0
